don't even know why I fuck with SFML/C++ anymore. + Simulate toggle + manual-labor-intensive fixed update/ pre update in TileScene2D. I want to manage the update in each scene rather than fuck about with the sf::Clock in the Main.cpp loop.
+ increase/decrease FPS... prerequisite the previous +. Big Weed.+--+'

+++ CellularAutomataCustom/Cell.cpp
@@ -15,7 +15,7 @@ Cell::~Cell()
 {
 }
 
-void Cell::Update(float a_DeltaTime)
+void Cell::Update(double a_DeltaTime)
 {
 	//first
 	m_IsAlive = m_WillBeAliveOnUpdate;
